requirements.txt
google-genai>=1.68.0